Move the `to_canonical_value` function to the `reference_context` module.
PiperOrigin-RevId: 358050176
