+++ tensorflow_federated/python/core/impl/BUILD
@@ -117,7 +117,6 @@ py_library(
     srcs_version = "PY3",
     deps = [
         "//tensorflow_federated/python/common_libs:py_typecheck",
-        "//tensorflow_federated/python/common_libs:structure",
         "//tensorflow_federated/python/core/api:computation_types",
         "//tensorflow_federated/python/core/api:typed_object",
     ],
@@ -131,7 +130,6 @@ py_test(
     srcs_version = "PY3",
     deps = [
         ":type_utils",
-        "//tensorflow_federated/python/common_libs:structure",
         "//tensorflow_federated/python/core/api:computation_types",
         "//tensorflow_federated/python/core/api:test_case",
         "//tensorflow_federated/python/core/impl/compiler:building_block_factory",
