+++ tensorflow_federated/python/core/impl/type_utils_test.py
@@ -12,12 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import collections
-
 from absl.testing import parameterized
 import tensorflow as tf
 
-from tensorflow_federated.python.common_libs import structure
 from tensorflow_federated.python.core.api import computation_types
 from tensorflow_federated.python.core.api import test_case
 from tensorflow_federated.python.core.impl import type_utils
@@ -25,96 +22,6 @@ from tensorflow_federated.python.core.impl.compiler import building_block_factor
 
 
 class TypeUtilsTest(test_case.TestCase, parameterized.TestCase):
-
-  def test_to_canonical_value_with_none(self):
-    self.assertIsNone(type_utils.to_canonical_value(None))
-
-  def test_to_canonical_value_with_int(self):
-    self.assertEqual(type_utils.to_canonical_value(1), 1)
-
-  def test_to_canonical_value_with_float(self):
-    self.assertEqual(type_utils.to_canonical_value(1.0), 1.0)
-
-  def test_to_canonical_value_with_bool(self):
-    self.assertEqual(type_utils.to_canonical_value(True), True)
-    self.assertEqual(type_utils.to_canonical_value(False), False)
-
-  def test_to_canonical_value_with_string(self):
-    self.assertEqual(type_utils.to_canonical_value('a'), 'a')
-
-  def test_to_canonical_value_with_list_of_ints(self):
-    self.assertEqual(type_utils.to_canonical_value([1, 2, 3]), [1, 2, 3])
-
-  def test_to_canonical_value_with_list_of_floats(self):
-    self.assertEqual(
-        type_utils.to_canonical_value([1.0, 2.0, 3.0]), [1.0, 2.0, 3.0])
-
-  def test_to_canonical_value_with_list_of_bools(self):
-    self.assertEqual(
-        type_utils.to_canonical_value([True, False]), [True, False])
-
-  def test_to_canonical_value_with_list_of_strings(self):
-    self.assertEqual(
-        type_utils.to_canonical_value(['a', 'b', 'c']), ['a', 'b', 'c'])
-
-  def test_to_canonical_value_with_list_of_dict(self):
-    self.assertEqual(
-        type_utils.to_canonical_value([{
-            'a': 1,
-            'b': 0.1,
-        }]), [structure.Struct([
-            ('a', 1),
-            ('b', 0.1),
-        ])])
-
-  def test_to_canonical_value_with_list_of_ordered_dict(self):
-    self.assertEqual(
-        type_utils.to_canonical_value(
-            [collections.OrderedDict([
-                ('a', 1),
-                ('b', 0.1),
-            ])]), [structure.Struct([
-                ('a', 1),
-                ('b', 0.1),
-            ])])
-
-  def test_to_canonical_value_with_dict(self):
-    self.assertEqual(
-        type_utils.to_canonical_value({
-            'a': 1,
-            'b': 0.1,
-        }), structure.Struct([
-            ('a', 1),
-            ('b', 0.1),
-        ]))
-    self.assertEqual(
-        type_utils.to_canonical_value({
-            'b': 0.1,
-            'a': 1,
-        }), structure.Struct([
-            ('a', 1),
-            ('b', 0.1),
-        ]))
-
-  def test_to_canonical_value_with_ordered_dict(self):
-    self.assertEqual(
-        type_utils.to_canonical_value(
-            collections.OrderedDict([
-                ('a', 1),
-                ('b', 0.1),
-            ])), structure.Struct([
-                ('a', 1),
-                ('b', 0.1),
-            ]))
-    self.assertEqual(
-        type_utils.to_canonical_value(
-            collections.OrderedDict([
-                ('b', 0.1),
-                ('a', 1),
-            ])), structure.Struct([
-                ('b', 0.1),
-                ('a', 1),
-            ]))
 
   # pyformat: disable
   @parameterized.named_parameters([
